nilfs2: fix arithmetic range overflow of alloc.h
The inline function nilfs_persistent_group_bitmap_blkoff cause an
arithmetic range overflow on multiplication.
This bug was imported by the commit
32c2fe67bb65db56b0c81420aae53049032648b2 which removes nilfs_bgno_t
type.

Acked-by: Ryusuke Konishi <konishi.ryusuke@lab.ntt.co.jp>

+++ fs/alloc.h
@@ -74,7 +74,7 @@ nilfs_persistent_group_desc_blkoff(struct inode *inode, unsigned long group)
 static inline unsigned long
 nilfs_persistent_group_bitmap_blkoff(struct inode *inode, unsigned long group)
 {
-	unsigned long group_offset =
+	unsigned long long group_offset =
 		group / nilfs_persistent_group_descs_per_block(inode);
 
 	return nilfs_persistent_group_desc_blkoff(inode, group) + 1 +
